web/requirements.txt
streamlit
pandas
streamlit-extras
firebase-admin
plotly.graph_objects
cohere